net: lib: download_client: skip duplicate packets
Duplicate packets are not handled/skipped that results in a continuous
write of the fota package to dfu.

The reason for this is that if recv fails due to a bad connection
(timeout etc), always the current block is requested. Once the
connection is back it is likely that all the requested blocks are
received.

Signed-off-by: Andreas Chmielewski <andreas.chmielewski@grandcentrix.net>
(cherry picked from commit a654fb6ef602b05f69877e16143bbe6900ad6e73)